+++ src/components/pages/QuestionListPage.jsx
@@ -4,11 +4,15 @@ import CardSortDropdown from '../atoms/Dropdown/CardSortDropdown';
 import UserCardList from '../organisms/UserCardList/UserCardList';
 import useQuestionListPageData from '../../hooks/useQuestionListPageData';
 import PageNationNumbersContainer from '../organisms/PageNationNumbersContainer/PageNationNumbersContainer';
-import PageNationButton from '../organisms/PageNationButton/PageNationButton';
+import PageNationButton from '../organisms/PageNationNumbersContainer/PageNationNumbers/PageNationButton/PageNationButton';
 import RandomCardButton from '../atoms/Button/RandomCardButton/RandomCardButton';
 import RandomQuestionCard from '../organisms/RandomQuestionCard/RandomQuestionCard';
 import useShowRandomCard from '../../hooks/useShowRandomCard';
 
+/**
+ *
+ * @returns QuestionListPage Component
+ */
 export default function QuestionListPage() {
   const {
     currentPage,
@@ -25,7 +29,7 @@ export default function QuestionListPage() {
   const {
     showRandomCard,
     handleShowRandomCardButtonClick,
-    handleModalCloseButtonClick,
+    handleRandomCardCloseClick,
   } = useShowRandomCard();
 
   return (
@@ -66,7 +70,7 @@ export default function QuestionListPage() {
       </PageNationButtonContainer>
       {/* 랜덤 질문 카드가 보이는 실험적인 기능입니다! */}
       <RandomQuestionCard
-        onModalCloseButtonClick={handleModalCloseButtonClick}
+        onRandomCardCloseClick={handleRandomCardCloseClick}
         showRandomCard={showRandomCard}
       />
     </Wrapper>
